+++ src/features/editor/extensions.ts
@@ -23,9 +23,9 @@ import { PrismCodeBlock } from "./extensions/prism-code-block";
 import { PrismCodeContent } from "./extensions/prism-code-block/prism-code-content";
 import { PrismCodeFileName } from "./extensions/prism-code-block/prism-code-file-name";
 import { Placeholder } from "./extensions/placeholder";
-import { EmbedLinkCard } from "./extensions/embed-link-card";
-import { EmbedTweet } from "./extensions/embed-tweet";
-import { EmbedGithub } from "./extensions/embed-github";
+import { EmbedLinkCard } from "./extensions/embed/embed-link-card";
+import { EmbedTweet } from "./extensions/embed/embed-tweet";
+import { EmbedGithub } from "./extensions/embed/embed-github";
 
 export const extensions: Extensions = [
   // === Core ===
